refactor: 기존 JDBC 관련 코드 백업 디렉토리 생성
