+++ src/main/java/com/example/bcsd/controller/dto/request/ArticleCreateRequest.java
@@ -1,6 +1,6 @@
 package com.example.bcsd.controller.dto.request;
 
-import com.example.bcsd.model.Article;
+import com.example.bcsd.domain.beforejpa.Article;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
