+++ src/main/java/com/example/bcsd/domain/Article.java
@@ -1,0 +1,4 @@
+package com.example.bcsd.domain;
+
+public class Article {
+}
