+++ src/main/java/com/example/bcsd/repository/beforejpa/BoardRepository.java
@@ -1,6 +1,6 @@
 package com.example.bcsd.repository;
 
-import com.example.bcsd.model.Board;
+import com.example.bcsd.domain.beforejpa.Board;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
