+++ src/main/java/com/example/bcsd/service/BoardService.java
@@ -5,7 +5,7 @@ import com.example.bcsd.controller.dto.request.BoardUpdateRequest;
 import com.example.bcsd.controller.dto.response.BoardResponse;
 import com.example.bcsd.global.exception.CustomException;
 import com.example.bcsd.global.exception.ErrorCode;
-import com.example.bcsd.model.Board;
+import com.example.bcsd.domain.beforejpa.Board;
 import com.example.bcsd.repository.BoardRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
